requirements.txt
aiocron==1.8
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.0.1
croniter==1.3.8
distlib==0.3.6
filelock==3.9.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
nextcord==2.3.3
platformdirs==2.6.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz-deprecation-shim==0.1.0.post0
schedule==1.1.0
six==1.16.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
virtualenv==20.17.1
yarl==1.8.2
